tokenize: handle Apertium-style superblanks in --giella-cg format
Also support Apertium-style flushing, triggered by NUL in input,
outputting a <STREAMCMD:FLUSH> since output is CG-format

Changes outside --giella-cg: push constness of input_text into
pmatch.cc::locate (doesn't change the string, so no need to be mut)

+tests

+++ test/tools/tokenize-functionality.sh
@@ -65,5 +65,16 @@ if ! diff test.strings $srcdir/tokenize-dog-out-xerox.strings ; then
 fi
 
 
+# --giella-cg superblanks
+if ! printf 'dog[\\\n<\\\\>]cat !and \ndogs[][\n]' | $TOOLDIR/hfst-tokenize --giella-cg $srcdir/tokenize-dog.pmhfst > test.strings ; then
+    echo tokenize --giella-cg superblank fail:
+    cat test.strings
+    exit 1
+fi
+if ! diff test.strings $srcdir/tokenize-dog-out-giella-cg-superblank.strings ; then
+    echo diff test.strings $srcdir/tokenize-dog-out-giella-cg-superblank.strings 
+    exit 1
+fi
+
 rm test.strings tokenize-dog.pmhfst tokenize-dog.hfst tokenize-dog-gen.hfst
 exit 0
